+++ src/app/parser/typescript/test-data/morph-tests/imports/fresh.ts
@@ -1,0 +1,5 @@
+import * as a from "./a";
+import * as tsMorph from "ts-morph";
+import { exit } from "process";
+import { SemVer } from "semver";
+